+++ imagesearch-20201214/alibabacloud_imagesearch20201214/models/_check_image_exists_request.py
@@ -1,0 +1,52 @@
+# -*- coding: utf-8 -*-
+# This file is auto-generated, don't edit it. Thanks.
+from __future__ import annotations
+
+from darabonba.model import DaraModel
+
+class CheckImageExistsRequest(DaraModel):
+    def __init__(
+        self,
+        instance_name: str = None,
+        pic_name: str = None,
+        product_id: str = None,
+    ):
+        # This parameter is required.
+        self.instance_name = instance_name
+        # This parameter is required.
+        self.pic_name = pic_name
+        # This parameter is required.
+        self.product_id = product_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.instance_name is not None:
+            result['InstanceName'] = self.instance_name
+
+        if self.pic_name is not None:
+            result['PicName'] = self.pic_name
+
+        if self.product_id is not None:
+            result['ProductId'] = self.product_id
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('InstanceName') is not None:
+            self.instance_name = m.get('InstanceName')
+
+        if m.get('PicName') is not None:
+            self.pic_name = m.get('PicName')
+
+        if m.get('ProductId') is not None:
+            self.product_id = m.get('ProductId')
+
+        return self
+